+++ src/main/java/org/jabref/migrations/PreferencesMigrations.java
@@ -55,6 +55,7 @@ public class PreferencesMigrations {
         upgradeColumnPreferences(Globals.prefs);
         restoreVariablesForBackwardCompatibility(Globals.prefs);
         upgradePreviewStyleAllowMarkdown(Globals.prefs);
+        migrateDoNotResolveFields(Globals.prefs);
     }
 
     /**
@@ -288,6 +289,18 @@ public class PreferencesMigrations {
         keys.add("related");
         keys.add("entryset");
         prefs.putStringList(JabRefPreferences.AUTOCOMPLETER_COMPLETE_FIELDS, new ArrayList<>(keys));
+    }
+
+    private static void migrateDoNotResolveFields(JabRefPreferences prefs) {
+        Set<String> keys = new LinkedHashSet<>(prefs.getStringList(JabRefPreferences.DO_NOT_RESOLVE_STRINGS_FOR));
+        keys.add("url");
+        keys.add("abstract");
+        keys.add("note");
+        keys.add("comment");
+        keys.add("doi");
+        keys.add("eprint");
+        keys.add("file");
+        prefs.putStringList(JabRefPreferences.DO_NOT_RESOLVE_STRINGS_FOR, new ArrayList<>(keys));
     }
 
     private static void migrateTypedKeyPrefs(JabRefPreferences prefs, Preferences oldPatternPrefs)